+++ apps/web/.env.local.example
@@ -25,6 +25,11 @@ NEXT_PUBLIC_API_URL=http://localhost:8000/api
 # NEXT_PUBLIC_API_URL=https://api.tripful.com/api
 # NEXT_PUBLIC_API_URL=https://your-domain.com/api
 
+# Server-side API URL (for RSC server-side data fetching)
+# This is NOT exposed to the browser - used only by Next.js server components.
+# In production, this may point to an internal/private API endpoint.
+API_URL=http://localhost:8000/api
+
 # ===========================================
 # Optional: Additional Frontend Configuration
 # ===========================================
