+++ apps/web/src/app/(app)/layout.tsx
@@ -1,29 +1,17 @@
-"use client";
+import { cookies } from "next/headers";
+import { redirect } from "next/navigation";
+import type { ReactNode } from "react";
 
-import { useEffect, type ReactNode } from "react";
-import { useRouter } from "next/navigation";
-import { useAuth } from "@/app/providers/auth-provider";
+export default async function ProtectedLayout({
+  children,
+}: {
+  children: ReactNode;
+}) {
+  const cookieStore = await cookies();
+  const authToken = cookieStore.get("auth_token");
 
-export default function ProtectedLayout({ children }: { children: ReactNode }) {
-  const { user, loading } = useAuth();
-  const router = useRouter();
-
-  useEffect(() => {
-    if (!loading && !user) {
-      router.push("/login");
-    }
-  }, [user, loading, router]);
-
-  if (loading) {
-    return (
-      <div className="flex items-center justify-center min-h-screen">
-        <div className="text-gray-600 animate-pulse">Loading...</div>
-      </div>
-    );
-  }
-
-  if (!user) {
-    return null;
+  if (!authToken?.value) {
+    redirect("/login");
   }
 
   return <>{children}</>;
